requirements.txt
# ===================================
# Geo-LLM Spatial RAG Engine
# Python Dependencies
# ===================================

# --- Phase 0: Data Acquisition ---
requests>=2.31.0

# --- Phase 1: Database & Ingestion ---
sqlalchemy>=2.0.0
geoalchemy2>=0.14.0
geopandas>=0.14.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
shapely>=2.0.0

# --- Phase 2: Backend API & AI ---
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
sqlparse>=0.5.0
sse-starlette>=2.0.0
